add support diy format

+++ library/src/main/java/com/cherry/lib/doc/office/pg/control/PGControl.java
@@ -600,7 +600,7 @@ public class PGControl extends AbstractControl
      * 
      *
      */
-    public byte getApplicationType()
+    public int getApplicationType()
     {
         return  MainConstant.APPLICATION_TYPE_PPT;
     }
