+++ library/src/main/java/com/cherry/lib/doc/office/system/IControl.java
@@ -111,7 +111,7 @@ public interface IControl
     /**
      * 
      */
-    public byte getApplicationType();
+    public int getApplicationType();
     
     /**
      *
