+++ library/src/main/java/com/cherry/lib/doc/office/wp/control/WPControl.java
@@ -568,7 +568,7 @@ public class WPControl extends AbstractControl
      * 
      *
      */
-    public byte getApplicationType()
+    public int getApplicationType()
     {
         return  MainConstant.APPLICATION_TYPE_WP;
     }
